requirements.txt
Django>=4.2
djangorestframework
gunicorn
python-json-logger
psycopg2-binary